Initial clean-out, add minting support to builder in full
